Clean up and added presets file path to list parsers output #1662
