+++ tests/cli/extraction_tool.py
@@ -8,6 +8,7 @@ import argparse
 import unittest
 
 from plaso.cli import extraction_tool
+from plaso.parsers import manager as parsers_manager
 
 from tests.cli import test_lib
 
@@ -33,6 +34,21 @@ class ExtractionToolTest(test_lib.CLIToolTestCase):
       '                        (defaults to 125000)',
       ''])
 
+  # TODO: add test for _CreateProcessingConfiguration
+
+  def testGetParserPresetsInformation(self):
+    """Tests the _GetParserPresetsInformation function."""
+    test_tool = extraction_tool.ExtractionTool()
+
+    presets_file = self._GetTestFilePath(['presets.yaml'])
+    parsers_manager.ParsersManager.ReadPresetsFromFile(presets_file)
+
+    parser_presets_information = test_tool._GetParserPresetsInformation()
+    self.assertGreaterEqual(len(parser_presets_information), 1)
+
+    available_parser_names = [name for name, _ in parser_presets_information]
+    self.assertIn('linux', available_parser_names)
+
   def testParsePerformanceOptions(self):
     """Tests the _ParsePerformanceOptions function."""
     test_tool = extraction_tool.ExtractionTool()
@@ -40,6 +56,12 @@ class ExtractionToolTest(test_lib.CLIToolTestCase):
     options = test_lib.TestOptions()
 
     test_tool._ParsePerformanceOptions(options)
+
+  # TODO: add test for _ParseProcessingOptions
+  # TODO: add test for _PreprocessSources
+  # TODO: add test for _ReadParserPresetsFromFile
+  # TODO: add test for _SetExtractionParsersAndPlugins
+  # TODO: add test for _SetExtractionPreferredTimeZone
 
   def testAddPerformanceOptions(self):
     """Tests the AddPerformanceOptions function."""
@@ -52,6 +74,65 @@ class ExtractionToolTest(test_lib.CLIToolTestCase):
 
     output = self._RunArgparseFormatHelp(argument_parser)
     self.assertEqual(output, self._EXPECTED_PERFORMANCE_OPTIONS)
+
+  # TODO: add test for AddProcessingOptions
+
+  def testListParsersAndPlugins(self):
+    """Tests the ListParsersAndPlugins function."""
+    output_writer = test_lib.TestOutputWriter(encoding='utf-8')
+    test_tool = extraction_tool.ExtractionTool(output_writer=output_writer)
+
+    presets_file = self._GetTestFilePath(['presets.yaml'])
+    parsers_manager.ParsersManager.ReadPresetsFromFile(presets_file)
+
+    test_tool.ListParsersAndPlugins()
+
+    output = output_writer.ReadOutput()
+
+    number_of_tables = 0
+    lines = []
+    for line in output.split('\n'):
+      line = line.strip()
+      lines.append(line)
+
+      if line.startswith('*****') and line.endswith('*****'):
+        number_of_tables += 1
+
+    self.assertIn('Parsers', lines[1])
+
+    lines = frozenset(lines)
+
+    self.assertEqual(number_of_tables, 10)
+
+    expected_line = 'filestat : Parser for file system stat information.'
+    self.assertIn(expected_line, lines)
+
+    expected_line = 'bencode_utorrent : Parser for uTorrent bencoded files.'
+    self.assertIn(expected_line, lines)
+
+    expected_line = (
+        'msie_webcache : Parser for MSIE WebCache ESE database files.')
+    self.assertIn(expected_line, lines)
+
+    expected_line = 'olecf_default : Parser for a generic OLECF item.'
+    self.assertIn(expected_line, lines)
+
+    expected_line = 'plist_default : Parser for plist files.'
+    self.assertIn(expected_line, lines)
+
+    # Note that the expected line is truncated by the cell wrapping in
+    # the table.
+    expected_line = (
+        'chrome_27_history : Parser for Google Chrome 27 and up history SQLite')
+    self.assertIn(expected_line, lines)
+
+    expected_line = 'ssh : Parser for SSH syslog entries.'
+    self.assertIn(expected_line, lines)
+
+    expected_line = 'winreg_default : Parser for Registry data.'
+    self.assertIn(expected_line, lines)
+
+
 
 
 if __name__ == '__main__':
